+++ examples/seek.rs
@@ -6,8 +6,7 @@ use tokio::io::AsyncWriteExt;
 async fn main() -> io::Result<()> {
     let file = File::create("tests/seek.txt")?;
     file.set_len(0x11)?;
-    let file = tokio_file_unix::File::new_nb(file)?;
-    let mut file = file.into_io()?;
+    let mut file = tokio_file_unix::File::new_nb(file)?;
 
     file.write_all(b"aaaaAAAAaaaaAAAA\n").await?;
     file.get_mut().seek(SeekFrom::Start(8))?;