+++ changelog.md
@@ -2,6 +2,8 @@
 
 ## 0.6.0
 
+  - `File::get_nonblocking` and `File::set_nonblocking` have been migrated to
+    module-level.
   - `StdFile` has been removed in favor of `raw_stdin`, `raw_stdout`, and
     `raw_stderr`.
   - `DelimCodec` has been removed in favor of `tokio_util::codec::FramedRead`.